dry up types
