+++ packages/backend/src/tokens/types.ts
@@ -1,3 +1,5 @@
+import type { MatchFunction } from '@clerk/shared/pathToRegexp';
+
 import type { ApiClient } from '../api';
 import type { VerifyTokenOptions } from './verify';
 
@@ -116,3 +118,16 @@ export type OrganizationSyncOptions = {
  * ```
  */
 type Pattern = string;
+
+export type OrganizationSyncTargetMatchers = {
+  OrganizationMatcher: MatchFunction<Partial<Record<string, string | string[]>>> | null;
+  PersonalAccountMatcher: MatchFunction<Partial<Record<string, string | string[]>>> | null;
+};
+
+/**
+ * Represents an organization or a personal account - e.g. an
+ * entity that can be activated by the handshake API.
+ */
+export type OrganizationSyncTarget =
+  | { type: 'personalAccount' }
+  | { type: 'organization'; organizationId?: string; organizationSlug?: string };
